requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# ML Models
prophet>=1.1.0
xgboost>=2.0.0
scikit-learn>=1.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Experiment Tracking
mlflow>=2.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.11.0
